+++ packages/base/src/components/Dropdown/index.ts
@@ -1,4 +1,3 @@
 export { default as Dropdown } from './Dropdown';
 export * from './Dropdown';
 export * from './Dropdown.types';
-export * from './Dropdown.styled';