+++ packages/base/src/components/index.ts
@@ -17,3 +17,4 @@ export * from './Tag';
 export * from './Icon';
 export * from './Inputs';
 export * from './ControlButtons';
+export * from './Tooltip';