+++ packages/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @shoplflow/base
 
+## 0.24.30
+
+### Patch Changes
+
+- [#246](https://github.com/shopl/shoplflow/pull/246) [`3ae6a28`](https://github.com/shopl/shoplflow/commit/3ae6a28389b731ef4bfe4a0ba17f42db4ea7f09c) Thanks [@velo-kim](https://github.com/velo-kim)! - Fix: getStyleByType(input.styyled.ts) 불필요한 조건문 제거
+
+## 0.24.29
+
+### Patch Changes
+
+- [#244](https://github.com/shopl/shoplflow/pull/244) [`21e0ebf`](https://github.com/shopl/shoplflow/commit/21e0ebf706922ba7af6e9d91930a5f87cebc3fe3) Thanks [@velo-kim](https://github.com/velo-kim)! - Tooltip 컴포넌트 추가 및 각 컴포넌트 index 파일 styled import 제거
+
 ## 0.24.28
 
 ### Patch Changes