+++ test-env/CHANGELOG.md
@@ -1,5 +1,19 @@
 # test-env
 
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies [[`3ae6a28`](https://github.com/shopl/shoplflow/commit/3ae6a28389b731ef4bfe4a0ba17f42db4ea7f09c)]:
+  - @shoplflow/base@0.24.30
+
+## 0.0.70
+
+### Patch Changes
+
+- Updated dependencies [[`21e0ebf`](https://github.com/shopl/shoplflow/commit/21e0ebf706922ba7af6e9d91930a5f87cebc3fe3)]:
+  - @shoplflow/base@0.24.29
+
 ## 0.0.69
 
 ### Patch Changes